int tests

+++ safe-road-backend/module-learning/src/main/java/ru/itmo/saferoad/learning/domain/repository/TestSessionRepository.java
@@ -3,5 +3,5 @@ package ru.itmo.saferoad.learning.domain.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itmo.saferoad.learning.domain.TestSession;
 
-public interface TestSessionRepository extends JpaRepository<TestSession, Integer> {
+public interface TestSessionRepository extends JpaRepository<TestSession, Long> {
 }
